Add support for a single URL argument to populate all options
Entering one by one all the properties that are basically all
the parts of a URI (even username:password could be specified
in a single value) is supper annoying. So make it possible
for all commands to also receive a URL argument (no need for
another option) that can populate all options automatically.

Fixes #2