Add backwards compatibility renaming for --protocol
Which is now --scheme.

To make the option consistent across all three commands, we renamed --protocol to --scheme so we could always use the shorthand -s regardless of whether there was a -p|--password option too.

This caused older invocations to break since the protocol would be null/empty and thus a matching provider was not found (i.e. for github).

Fixes #25